Huge commit! Combined both the backend and frontend into one single application by some clever threading. No need to communicate with the other class or make shared data structures.
